+++ server/routes/index.js
@@ -20,6 +20,28 @@ const {
   getDashboardStatistics,
   getCurrentOperations
 } = require("../services/printer-statistics.service");
+const { SystemRunner } = require("../services/system-information.service");
+const { fetchUsers } = require("../services/user-service");
+const { fetchMongoDBConnectionString, fetchClientVersion } = require("../app-env");
+const isDocker = require("is-docker");
+const { isNodemon, isNode, isPm2 } = require("../utils/env.utils");
+const { getCurrentBranch, checkIfWereInAGitRepo } = require("../utils/git.utils");
+const { returnPatreonData } = require("../services/patreon.service");
+const fs = require("fs");
+const marked = require("marked");
+
+marked.setOptions({
+  renderer: new marked.Renderer(),
+  smartLists: true,
+  smartypants: true
+});
+
+const md = function (filename) {
+  const path = "../" + filename;
+  const include = fs.readFileSync(path, "utf8");
+  const html = marked.parse(include);
+  return html;
+};
 
 const version = process.env[AppConstants.VERSION_KEY];
 
@@ -319,6 +341,45 @@ router.get("/filament", ensureAuthenticated, ensureCurrentUserAndGroup, async (r
     statistics,
     historyStats,
     clientSettings: req.user.clientSettings
+  });
+});
+
+router.get("/system", ensureAuthenticated, ensureCurrentUserAndGroup, async (req, res) => {
+  const clientSettings = await SettingsClean.returnClientSettings();
+  const serverSettings = SettingsClean.returnSystemSettings();
+  const systemInformation = SystemRunner.returnInfo();
+  const softwareUpdateNotification = softwareUpdateChecker.getUpdateNotificationIfAny();
+  let dashboardSettings = clientSettings?.dashboard || getDefaultDashboardSettings();
+  const currentUsers = await fetchUsers();
+
+  res.render("system", {
+    name: req.user.name,
+    userGroup: req.user.group,
+    version: process.env[AppConstants.VERSION_KEY],
+    printerCount: getPrinterStoreCache().getPrinterCount(),
+    page: "System",
+    octoFarmPageTitle: process.env[AppConstants.OCTOFARM_SITE_TITLE_KEY],
+    helpers: prettyHelpers,
+    clientSettings,
+    serverSettings,
+    systemInformation,
+    md,
+    db: fetchMongoDBConnectionString(),
+    dashboardSettings: dashboardSettings,
+    serviceInformation: {
+      isDockerContainer: isDocker(),
+      isNodemon: isNodemon(),
+      isNode: isNode(),
+      isPm2: isPm2(),
+      update: softwareUpdateNotification
+    },
+    currentGitBranch: await getCurrentBranch(),
+    clientVersion: fetchClientVersion(),
+    areWeGitRepo: await checkIfWereInAGitRepo(),
+    systemEnvironment: process.env[AppConstants.NODE_ENV_KEY],
+    patreonData: returnPatreonData(),
+    currentUsers,
+    taskManagerState: TaskManager.getTaskState()
   });
 });